Major pipeline writer refactor
