+++ lineapy/plugins/base_pipeline_writer.py
@@ -7,11 +7,7 @@ from typing import Dict, List, Optional, Union
 
 import networkx as nx
 
-from lineapy.data.types import PipelineType
-from lineapy.graph_reader.artifact_collection import (
-    ArtifactCollection,
-    SessionArtifacts,
-)
+from lineapy.graph_reader.artifact_collection import ArtifactCollection
 from lineapy.graph_reader.node_collection import (
     ArtifactNodeCollection,
     UserCodeNodeCollection,
@@ -20,11 +16,8 @@ from lineapy.graph_reader.types import InputVariable
 from lineapy.plugins.session_writers import BaseSessionWriter
 from lineapy.plugins.task import (
     AirflowDagConfig,
-    AirflowDagFlavor,
     DVCDagConfig,
-    DVCDagFlavor,
     TaskDefinition,
-    TaskGraph,
     TaskGraphEdge,
 )
 from lineapy.plugins.utils import (
@@ -402,244 +395,29 @@ class BasePipelineWriter:
         self._write_docker()
         self._create_test()
 
-
-class AirflowPipelineWriter(BasePipelineWriter):
-    """
-    Class for pipeline file writer. Corresponds to "AIRFLOW" framework.
-    """
-
-    @property
-    def docker_template_name(self) -> str:
-        return "airflow_dockerfile.jinja"
-
-    def _write_dag(self) -> None:
-        dag_flavor = self.dag_config.get(
-            "dag_flavor", "PythonOperatorPerArtifact"
-        )
-
-        # Check if the given DAG flavor is a supported/valid one
-        if dag_flavor not in AirflowDagFlavor.__members__:
-            raise ValueError(
-                f'"{dag_flavor}" is an invalid airflow dag flavor.'
-            )
-
-        # Construct DAG text for the given flavor
-        if (
-            AirflowDagFlavor[dag_flavor]
-            == AirflowDagFlavor.PythonOperatorPerSession
-        ):
-            full_code = self._write_operator_per_session()
-        elif (
-            AirflowDagFlavor[dag_flavor]
-            == AirflowDagFlavor.PythonOperatorPerArtifact
-        ):
-            full_code = self._write_operator_per_artifact()
-
-        # Write out file
-        file = self.output_dir / f"{self.pipeline_name}_dag.py"
-        file.write_text(prettify(full_code))
-        logger.info(f"Generated DAG file: {file}")
-
-    def _write_operator_per_session(self) -> str:
-        """
-        This hidden method implements Airflow DAG code generation corresponding
-        to the `PythonOperatorPerSession` flavor, where each session gets its
-        own Python operator. For instance, if the two artifacts in our pipeline
-        (e.g., model and prediction) were created in the same session, we would
-        get an Airflow DAG file looking as the following:
-
-        .. code-block:: python
-        def dag_setup():
-            pickle_folder = pathlib.Path("/tmp").joinpath("g2_z")
-            if not pickle_folder.exists():
-                pickle_folder.mkdir()
-
-
-        def dag_teardown():
-            pickle_files = pathlib.Path("/tmp").joinpath("g2_z").glob("*.pickle")
-            for f in pickle_files:
-                f.unlink()
-
-            def task_run_session_including_g2():
-                artifacts = g2_z_module.run_session_including_g2()
-                pickle.dump(artifacts["g2"], open("/tmp/g2_z/artifact_g2.pickle", "wb"))
-                pickle.dump(artifacts["z"], open("/tmp/g2_z/artifact_z.pickle", "wb"))
-
-            with DAG(...) as dag:
-                setup = PythonOperator(
-                    task_id="dag_setup",
-                    python_callable=dag_setup,
-                )
-
-                teardown = PythonOperator(
-                    task_id="dag_teardown",
-                    python_callable=dag_teardown,
-                )
-
-                run_session_including_g2 = PythonOperator(
-                    task_id="run_session_including_g2_task",
-                    python_callable=task_run_session_including_g2,
-                )
-
-                setup >> run_session_including_g2
-                run_session_including_g2 >> teardown
-
-            run_session_including_g2
-
-        This way, the generated Airflow DAG file opens room for engineers
-        to peak and control pipeline runs at a finer level and allows
-        for further customization.
-        """
-        codegenerator = AirflowCodeGenerator(self.artifact_collection)
-        DAG_TEMPLATE = load_plugin_template("airflow_dag_PythonOperator.jinja")
-        task_functions = []
-        task_definitions = []
+    def get_pipeline_args(self) -> Dict[str, InputVariable]:
+        pipeline_args: Dict[str, InputVariable] = dict()
         for sa in self.session_artifacts_sorted:
-            task_functions.append(
-                BaseSessionWriter().get_session_function_name(sa)
+            pipeline_args.update(
+                BaseSessionWriter().get_session_input_parameters_spec(sa)
             )
-            task_definitions.append(
-                get_session_task_definition(sa, self.pipeline_name)
-            )
-        dependencies = {
-            task_functions[i + 1]: {task_functions[i]}
-            for i in range(len(task_functions) - 1)
-        }
-        task_graph = TaskGraph(
-            nodes=task_functions,
-            mapping={f: f for f in task_functions},
-            edges=dependencies,
-        )
-        session_function_params = (
-            codegenerator.get_session_function_params_args()
-        )
-        tasks = [
-            {"name": ft, "op_kwargs": session_function_params.get(ft, None)}
-            for ft in task_functions
-        ]
-        full_code = DAG_TEMPLATE.render(
-            DAG_NAME=self.pipeline_name,
-            MODULE_NAME=self.pipeline_name + "_module",
-            OWNER=self.dag_config.get("owner", "airflow"),
-            RETRIES=self.dag_config.get("retries", 2),
-            START_DATE=self.dag_config.get("start_date", "days_ago(1)"),
-            SCHEDULE_INTERVAL=self.dag_config.get(
-                "schedule_interval", "*/15 * * * *"
-            ),
-            MAX_ACTIVE_RUNS=self.dag_config.get("max_active_runs", 1),
-            CATCHUP=self.dag_config.get("catchup", "False"),
-            dag_params=codegenerator.get_params_args(),
-            task_definitions=task_definitions,
-            tasks=tasks,
-            task_dependencies=task_graph.get_airflow_dependencies(
-                setup_task="setup", teardown_task="teardown"
-            ),
-        )
-
-        return full_code
-
-    def _write_operator_per_artifact(self) -> str:
-        """
-        This method implements Airflow DAG code generation corresponding to the
-        `PythonOperatorPerArtifact` flavor, where each artifact gets its own
-        Python operator. For instance, if the two artifacts in our pipeline
-        (e.g., model and prediction) were created in the same session, we would
-        get an Airflow DAG file looking as the following:
-
-        .. code-block:: python
-            import pickle
-            import iris_module
-
-            ...
-
-            def task_iris_model():
-                mod = iris_module.get_iris_model()
-                pickle.dump(mod, open("/tmp/iris/variable_mod.pickle", "wb"))
-
-            def task_iris_pred():
-                mod = pickle.load(open("/tmp/iris/variable_mod.pickle", "rb"))
-                pred = iris_module.get_iris_pred(mod)
-                pickle.dump(
-                    pred, open("/tmp/iris/variable_pred.pickle", "wb")
-                )
-
-            with DAG(...) as dag:
-                iris_model = PythonOperator(
-                    task_id="iris_model_task",
-                    python_callable=task_iris_model,
-                )
-
-                iris_pred = PythonOperator(
-                    task_id="iris_pred_task",
-                    python_callable=task_iris_pred,
-                )
-
-            iris_model >> iris_pred
-
-        This way, the generated Airflow DAG file opens room for engineers
-        to peak and control pipeline runs at a finer level and allows
-        for further customization.
-        """
-        codegenerator = AirflowCodeGenerator(self.artifact_collection)
-        DAG_TEMPLATE = load_plugin_template("airflow_dag_PythonOperator.jinja")
-        task_def = self.get_artifact_task_definitions()
-        task_functions = list(task_def.keys())
-        task_definitions = [task["definition"] for task in task_def.values()]
-        dependencies = {
-            task_functions[i + 1]: {task_functions[i]}
-            for i in range(len(task_functions) - 1)
-        }
-        task_graph = TaskGraph(
-            nodes=task_functions,
-            mapping={f: f for f in task_functions},
-            edges=dependencies,
-        )
-        artifact_function_params = (
-            codegenerator.get_artifact_function_params_args(task_def)
-        )
-        tasks = [
-            {"name": ft, "op_kwargs": artifact_function_params.get(ft, None)}
-            for ft in task_functions
-        ]
-        full_code = DAG_TEMPLATE.render(
-            DAG_NAME=self.pipeline_name,
-            MODULE_NAME=self.pipeline_name + "_module",
-            OWNER=self.dag_config.get("owner", "airflow"),
-            RETRIES=self.dag_config.get("retries", 2),
-            START_DATE=self.dag_config.get("start_date", "days_ago(1)"),
-            SCHEDULE_INTERVAL=self.dag_config.get(
-                "schedule_interval", "*/15 * * * *"
-            ),
-            dag_params=codegenerator.get_params_args(),
-            MAX_ACTIVE_RUNS=self.dag_config.get("max_active_runs", 1),
-            CATCHUP=self.dag_config.get("catchup", "False"),
-            task_definitions=task_definitions,
-            tasks=tasks,
-            task_dependencies=task_graph.get_airflow_dependencies(
-                setup_task="setup", teardown_task="teardown"
-            ),
-        )
-
-        return full_code
+        return pipeline_args
 
     def get_artifact_task_definitions(
-        self, indentation=4
+        self,
     ) -> Dict[str, TaskDefinition]:
         """
         Add deserialization of input variables and serialization of output
-        variables logic of the artifact fucntion call_block and wrap them into a
+        variables logic of the artifact function call_block and wrap them into a
         new function definition.
         """
         task_definitions: Dict[str, TaskDefinition] = dict()
         unused_input_parameters = set(
             self.artifact_collection.input_parameters
         )
+        session_input_parameters_spec = self.get_pipeline_args()
+
         for session_artifacts in self.session_artifacts_sorted:
-            session_input_parameters_spec = (
-                BaseSessionWriter().get_session_input_parameters_spec(
-                    session_artifacts
-                )
-            )
             for nc in session_artifacts.usercode_nodecollections:
                 all_input_variables = sorted(list(nc.input_variables))
                 artifact_user_input_variables = [
@@ -670,176 +448,72 @@ class AirflowPipelineWriter(BasePipelineWriter):
                     f"pickle.dump({var},open('/tmp/{self.pipeline_name}/variable_{var}.pickle','wb'))"
                     for var in nc.return_variables
                 ]
-                TASK_FUNCTION_TEMPLATE = load_plugin_template(
-                    "task_function.jinja"
-                )
-                function_definition = TASK_FUNCTION_TEMPLATE.render(
+
+                # function_name = pn
+                # user_input_variables = n, p
+                # typing_blocks = [n = int(n), p = str(p)]
+                # loading_block = # pickle.loads
+                # call_block = pn = airflow_pipeline_two_input_parameter_module.get_pn(n, p)
+                # dumping_blocks =
+                #     pickle.dump(
+                #         pn, open("/tmp/airflow_pipeline_two_input_parameter/variable_pn.pickle", "wb")
+                #     )
+
+                task_def: TaskDefinition = TaskDefinition(
                     function_name=nc.safename,
-                    user_input_variables=", ".join(
-                        artifact_user_input_variables
-                    ),
+                    user_input_variables=artifact_user_input_variables,
                     typing_blocks=user_input_var_typing_block,
                     loading_blocks=input_var_loading_block,
                     call_block=function_call_block,
                     dumping_blocks=return_var_saving_block,
-                    indentation_block=" " * indentation,
                 )
-                task_def: TaskDefinition = {
-                    "definition": function_definition,
-                    "user_input_variables": artifact_user_input_variables,
-                }
                 task_definitions[nc.safename] = task_def
 
         return task_definitions
 
+    def get_session_task_definition(
+        self,
+    ) -> Dict[str, TaskDefinition]:
+        """
+        Add serialization of output artifacts logic of the session function
+        call_block and wrap them into a new function definition.
+        """
+        task_definitions: Dict[str, TaskDefinition] = dict()
 
-def get_session_task_definition(
-    sa: SessionArtifacts,
-    pipeline_name: str,
-    indentation=4,
-) -> str:
-    """
-    Add serialization of output artifacts logic of the session function
-    call_block and wrap them into a new function definition.
-    """
-    session_input_parameters_spec = (
-        BaseSessionWriter().get_session_input_parameters_spec(sa)
-    )
-    session_input_variables = list(session_input_parameters_spec.keys())
-    user_input_var_typing_block = [
-        f"{var} = {session_input_parameters_spec[var].value_type}({var})"
-        for var in session_input_variables
-    ]
+        for session_artifacts in self.session_artifacts_sorted:
 
-    input_var_loading_block: List[str] = []
-    function_call_block = f"artifacts = {pipeline_name}_module.{BaseSessionWriter().get_session_function_callblock(sa)}"
-    return_artifacts_saving_block = [
-        f"pickle.dump(artifacts['{nc.name}'],open('/tmp/{pipeline_name}/artifact_{nc.safename}.pickle','wb'))"
-        for nc in sa.usercode_nodecollections
-        if isinstance(nc, ArtifactNodeCollection)
-    ]
-
-    TASK_FUNCTION_TEMPLATE = load_plugin_template("task_function.jinja")
-    return TASK_FUNCTION_TEMPLATE.render(
-        function_name=BaseSessionWriter().get_session_function_name(sa),
-        user_input_variables=", ".join(session_input_variables),
-        typing_blocks=user_input_var_typing_block,
-        loading_blocks=input_var_loading_block,
-        call_block=function_call_block,
-        dumping_blocks=return_artifacts_saving_block,
-        indentation_block=" " * indentation,
-    )
-
-
-class AirflowCodeGenerator:
-    def __init__(self, ac: ArtifactCollection) -> None:
-        self.artifact_collection = ac
-
-    def get_params_args(self) -> str:
-        input_parameters_dict = dict()
-        for sa in self.artifact_collection.sort_session_artifacts():
-            for input_spec in (
-                BaseSessionWriter()
-                .get_session_input_parameters_spec(sa)
-                .values()
-            ):
-                input_parameters_dict[
-                    input_spec.variable_name
-                ] = input_spec.value
-        return '"params":' + str(input_parameters_dict)
-
-    def get_session_function_params_args(self) -> Dict[str, str]:
-        session_function_input_parameters = dict()
-        for sa in self.artifact_collection.sort_session_artifacts():
-            session_input_parameters = list(sa.input_parameters_node.keys())
-            if len(session_input_parameters) > 0:
-                session_function_input_parameters[
-                    BaseSessionWriter().get_session_function_name(sa)
-                ] = "op_kwargs=" + str(
-                    {
-                        var: "{{ params." + var + " }}"
-                        for var in session_input_parameters
-                    }
+            session_input_parameters_spec = (
+                BaseSessionWriter().get_session_input_parameters_spec(
+                    session_artifacts
                 )
-        return session_function_input_parameters
+            )
+            session_input_variables = list(
+                session_input_parameters_spec.keys()
+            )
+            user_input_var_typing_block = [
+                f"{var} = {session_input_parameters_spec[var].value_type}({var})"
+                for var in session_input_variables
+            ]
 
-    def get_artifact_function_params_args(
-        self, artifact_function_definitions: Dict[str, TaskDefinition]
-    ) -> Dict[str, str]:
-        artifact_function_input_parameters = dict()
-        for sa in self.artifact_collection.sort_session_artifacts():
-            session_input_parameters = set(sa.input_parameters_node.keys())
-            for nc in sa.usercode_nodecollections:
-                user_input_variables = artifact_function_definitions[
-                    nc.safename
-                ]["user_input_variables"]
-                parameterized_variables = nc.input_variables.intersection(
-                    session_input_parameters
-                ).intersection(set(user_input_variables))
-                if len(parameterized_variables) > 0:
-                    artifact_function_input_parameters[
-                        nc.safename
-                    ] = "op_kwargs=" + str(
-                        {
-                            var: "{{ params." + var + " }}"
-                            for var in sorted(list(parameterized_variables))
-                        }
-                    )
-        return artifact_function_input_parameters
+            input_var_loading_block: List[str] = []
+            function_call_block = f"artifacts = {self.pipeline_name}_module.{BaseSessionWriter().get_session_function_callblock(session_artifacts)}"
+            return_artifacts_saving_block = [
+                f"pickle.dump(artifacts['{nc.name}'],open('/tmp/{self.pipeline_name}/artifact_{nc.safename}.pickle','wb'))"
+                for nc in session_artifacts.usercode_nodecollections
+                if isinstance(nc, ArtifactNodeCollection)
+            ]
+            function_name = BaseSessionWriter().get_session_function_name(
+                session_artifacts
+            )
+            task_def: TaskDefinition = TaskDefinition(
+                function_name=function_name,
+                user_input_variables=session_input_variables,
+                typing_blocks=user_input_var_typing_block,
+                loading_blocks=input_var_loading_block,
+                call_block=function_call_block,
+                dumping_blocks=return_artifacts_saving_block,
+            )
 
+            task_definitions[function_name] = task_def
 
-class DVCPipelineWriter(BasePipelineWriter):
-    def _write_dag(self) -> None:
-        dag_flavor = self.dag_config.get(
-            "dag_flavor", "SingleStageAllSessions"
-        )
-
-        # Check if the given DAG flavor is a supported/valid one
-        if dag_flavor not in DVCDagFlavor.__members__:
-            raise ValueError(f'"{dag_flavor}" is an invalid dvc dag flavor.')
-
-        # Construct DAG text for the given flavor
-        if DVCDagFlavor[dag_flavor] == DVCDagFlavor.SingleStageAllSessions:
-            full_code = self._write_operator_run_all_sessions()
-
-        # Write out file
-        file = self.output_dir / "dvc.yaml"
-        file.write_text(full_code)
-        logger.info(f"Generated DAG file: {file}")
-
-    def _write_operator_run_all_sessions(self) -> str:
-        """
-        This hidden method implements DVC DAG code generation corresponding
-        to the `SingleStageAllSessions` flavor. This DAG only has one stage and
-        calls `run_all_sessions` generated by the module file.
-        """
-
-        DAG_TEMPLATE = load_plugin_template(
-            "dvc_dag_SingleStageAllSessions.jinja"
-        )
-
-        full_code = DAG_TEMPLATE.render(
-            MODULE_COMMAND=f"python {self.pipeline_name}_module.py",
-        )
-
-        return full_code
-
-    @property
-    def docker_template_name(self) -> str:
-        return "dvc_dockerfile.jinja"
-
-
-class PipelineWriterFactory:
-    @classmethod
-    def get(
-        cls,
-        pipeline_type: PipelineType = PipelineType.SCRIPT,
-        *args,
-        **kwargs,
-    ):
-        if pipeline_type == PipelineType.AIRFLOW:
-            return AirflowPipelineWriter(*args, **kwargs)
-        elif pipeline_type == PipelineType.DVC:
-            return DVCPipelineWriter(*args, **kwargs)
-        else:
-            return BasePipelineWriter(*args, **kwargs)
+        return task_definitions
