+++ tests/unit/graph_reader/test_artifact_collection.py
@@ -5,8 +5,8 @@ import pytest
 
 from lineapy.api.models.linea_artifact import get_lineaartifactdef
 from lineapy.graph_reader.artifact_collection import ArtifactCollection
+from lineapy.plugins.base_pipeline_writer import BasePipelineWriter
 from lineapy.plugins.loader import load_as_module
-from lineapy.plugins.pipeline_writers import BasePipelineWriter
 
 
 @pytest.mark.parametrize(
